+++ irodsbackend/urls.py
@@ -7,13 +7,23 @@ app_name = 'irodsbackend'
 
 urlpatterns = [
     url(
-        regex=r'^stats/(?P<project>[0-9a-f-]+)\?path=(?P<path>[\w\-_/@]+)$',
-        view=views.IrodsStatisticsGetAPIView.as_view(),
+        regex=r'^api/stats/(?P<project>[0-9a-f-]+)\?path=(?P<path>[\w\-_/@]+)$',
+        view=views.IrodsStatisticsAPIView.as_view(),
         name='stats',
     ),
     url(
-        regex=r'^stats\?path=(?P<path>[\w\-_/@]+)$',
-        view=views.IrodsStatisticsGetAPIView.as_view(),
+        regex=r'^api/stats\?path=(?P<path>[\w\-_/@]+)$',
+        view=views.IrodsStatisticsAPIView.as_view(),
         name='stats',
+    ),
+    url(
+        regex=r'^api/list/(?P<project>[0-9a-f-]+)\?path=(?P<path>[\w\-_/@]+)$',
+        view=views.IrodsObjectListAPIView.as_view(),
+        name='list',
+    ),
+    url(
+        regex=r'^api/list\?path=(?P<path>[\w\-_/@]+)$',
+        view=views.IrodsObjectListAPIView.as_view(),
+        name='list',
     ),
 ]