+++ samplesheets/urls.py
@@ -46,12 +46,6 @@ urlpatterns = [
         view=views.SampleSheetDeleteView.as_view(),
         name='delete',
     ),
-    # Javascript API views
-    url(
-        regex=r'^irods/list/assay/(?P<assay>[0-9a-f-]+)?path=(?P<path>[\w\-_/]+)$',
-        view=views.IrodsObjectListAPIView.as_view(),
-        name='irods_list',
-    ),
     # Taskflow API views
     url(
         regex=r'^taskflow/dirs/get$',